Move dag package so we can reuse it
