+++ internal/dag/dag.go
