+++ internal/dag/ops.go
