+++ internal/dag/tarjan.go
