+++ internal/dag/walk.go
